pyproject.toml
[tool.poetry]
name = "brokeback_mountain_alleycat"
version = "0.1.0"
description = ""
authors = ["G"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.21"
python-dotenv = "^0.20.0"
Django = "^4.1"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
django-extensions = "^3.2.0"
django-richtextfield = "^1.6.1"
django-ckeditor = "^6.5.0"
django-solo = "^2.0.0"
requests = "^2.28.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
